Moving files to correct place in check folder structure
